move label binarizer to utils
