+++ grants_tagger_light/preprocessing/preprocess_wellcome.py
@@ -15,8 +15,11 @@ import pandas as pd
 import typer
 import yaml
 
-from grants_tagger_light.label_binarizer import create_label_binarizer
-from grants_tagger_light.utils import verify_if_paths_exist, split_data
+from grants_tagger_light.utils import (
+    verify_if_paths_exist,
+    split_data,
+    create_label_binarizer,
+)
 
 
 def process_tagger_data(tagger_data):
