+++ grants_tagger_light/utils/__init__.py
@@ -1,4 +1,10 @@
-from .utils import load_data, load_train_test_data, verify_if_paths_exist, write_jsonl
+from .utils import (
+    load_data,
+    load_train_test_data,
+    verify_if_paths_exist,
+    write_jsonl,
+    create_label_binarizer,
+)
 from .split_data import split_data
 
 __all__ = [
@@ -7,4 +13,5 @@ __all__ = [
     "verify_if_paths_exist",
     "write_jsonl",
     "split_data",
+    "create_label_binarizer",
 ]
